requirements.txt
requests==2.30.0
lxml==4.9.2
Pillow==9.5.0
PyYAML==6.0
loguru==0.7.0
cssselect==1.2.0
